requirements.txt
fastapi==0.136.0
uvicorn==0.44.0
httpx==0.28.1
groq==1.2.0
python-dotenv==1.2.2
python-jose[cryptography]==3.3.0
pydantic==2.13.1
